Card contentful

See merge request bitdarlene/darlene-backend!24

+++ backend/activities/routes.py
@@ -1,8 +1,8 @@
 from flask import (Blueprint, request)
 from flask_restful import Resource
 from backend import api, db
-from backend.activities.schemas import activities_schema
-from backend.activities.utils import create_activity, edit_activity
+from backend.activities.schemas import activity_schema, activities_schema
+from backend.activities.utils import create_activity, delete_cards, edit_activity
 from backend.models import Activity
 
 # Blueprint for activities
@@ -30,7 +30,12 @@ class ActivityCRUD(Resource):
     # Function to edit an activity
     def put(self):
         contentful_data = request.get_json()
-        edit_activity(contentful_data)
+        activity = Activity.query.filter_by(contentful_id=contentful_data["entityId"]).first()
+
+        if not activity:
+            return {"message": "Activity does not exist"}, 404
+
+        edit_activity(activity, contentful_data)
 
         db.session.commit()
 
@@ -44,13 +49,29 @@ class ActivityDelete(Resource):
         contentful_data = request.get_json()
         activity = Activity.query.filter_by(contentful_id=contentful_data["entityId"]).first()
 
+        if not activity:
+            return {"message": "Activity does not exist"}, 404
+
+        delete_cards(activity.cards)
+
         db.session.delete(activity)
         db.session.commit()
 
         return {"message": "Activity successfully deleted"}, 200
 
 
+# This class is used to get a specific activity based on id
+class ActivityGetSpecific(Resource):
+    def get(self, activity_id):
+        activity = Activity.query.get(activity_id)
+
+        if not activity:
+            return {"message": "Activity does not exist"}, 404
+
+        return activity_schema.dump(activity)
+
+
 # Creates the routes for the classes
 api.add_resource(ActivityCRUD, "/activities")
 api.add_resource(ActivityDelete, "/activities/delete")
-
+api.add_resource(ActivityGetSpecific, "/activities/<int:activity_id>")
